requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
bcrypt==4.0.1
    # via
    #   cadastro-vagas (pyproject.toml)
    #   passlib
blinker==1.9.0
    # via streamlit
cachetools==7.0.5
    # via streamlit
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.5
    # via requests
click==8.3.1
    # via streamlit
cryptography==46.0.5
    # via google-auth
distro==1.9.0
    # via google-genai
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-auth==2.49.1
    # via google-genai
google-genai==1.67.0
    # via cadastro-vagas (pyproject.toml)
greenlet==3.3.2
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   ollama
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.18.0
    # via
    #   altair
    #   plotly
numpy==2.4.3
    # via
    #   pandas
    #   pydeck
    #   streamlit
ollama==0.6.1
    # via cadastro-vagas (pyproject.toml)
packaging==26.0
    # via
    #   altair
    #   plotly
    #   streamlit
pandas==2.3.3
    # via
    #   cadastro-vagas (pyproject.toml)
    #   streamlit
passlib==1.7.4
    # via cadastro-vagas (pyproject.toml)
pillow==12.1.1
    # via streamlit
plotly==6.6.0
    # via cadastro-vagas (pyproject.toml)
protobuf==5.29.6
    # via streamlit
psycopg2-binary==2.9.11
    # via cadastro-vagas (pyproject.toml)
pyarrow==23.0.1
    # via streamlit
pyasn1==0.6.2
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   cadastro-vagas (pyproject.toml)
    #   google-genai
    #   ollama
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via cadastro-vagas (pyproject.toml)
pytz==2026.1.post1
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   cadastro-vagas (pyproject.toml)
    #   google-auth
    #   google-genai
    #   streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.3
    # via gitdb
sniffio==1.3.1
    # via google-genai
sqlalchemy==2.0.48
    # via cadastro-vagas (pyproject.toml)
streamlit==1.55.0
    # via cadastro-vagas (pyproject.toml)
tenacity==9.1.4
    # via
    #   google-genai
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.5
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
watchdog==6.0.0
    # via streamlit
websockets==16.0
    # via google-genai
